fix: refactor circular dependency fix fix: katana contract test case assertion
