+++ packages/core/src/contracts/market-gateway/MarketGatewayMultisendContract.ts
@@ -5,7 +5,7 @@ import { getConfig } from '../../configs';
 import { Token } from '../../services/tokens/data';
 import { getPaymentTokens } from '../../services/tokens/getPaymentTokens';
 import { ChainId, Erc721Order, Erc1155Order, WalletClient } from '../../types';
-import { createErc1155MarketGatewayContract, createMarketGatewayContract } from '../contract-creators';
+import { createErc1155MarketGatewayContract, createMarketGatewayContract } from './factory';
 import { InteractWithParamsStruct } from '../abis/types/v5/MarketGatewayMultisend';
 import { SettleParameterStruct } from '../abis/types/v5/MavisExchange';
 import { Erc1155MarketGatewayContract, SettleOrderData } from './Erc1155MarketGatewayContract';
