+++ packages/core/src/contracts/market-gateway/index.ts
@@ -1,3 +1,4 @@
 export * from './Erc1155MarketGatewayContract';
 export * from './MarketGatewayContract';
 export * from './MarketGatewayMultisendContract';
+export * from './factory';
